test: split PDF deferred float ordering cases
